@@ -4,38 +4,37 @@ import 'dart:io';
 import 'dart:typed_data';
 import 'package:flutter_midi_command_linux/flutter_midi_command_linux.dart';
 import 'package:flutter_midi_command_platform_interface/flutter_midi_command_platform_interface.dart';
-export 'package:flutter_midi_command_platform_interface/flutter_midi_command_platform_interface.dart'
-    show MidiDevice, MidiPacket, MidiPort;
+export 'package:flutter_midi_command_platform_interface/flutter_midi_command_platform_interface.dart' show MidiDevice, MidiPacket, MidiPort;
 
 class MidiCommand {
   factory MidiCommand() {
     if (_instance == null) {
       _instance = MidiCommand._();
     }
-    return _instance;
+    return _instance!;
   }
 
   MidiCommand._();
 
-  static MidiCommand _instance;
+  static MidiCommand? _instance;
 
-  static MidiCommandPlatform __platform;
+  static MidiCommandPlatform? __platform;
 
   StreamController<Uint8List> _txStreamCtrl = StreamController.broadcast();
 
   static MidiCommandPlatform get _platform {
-    if (__platform != null) return __platform;
+    if (__platform != null) return __platform!;
 
     if (Platform.isLinux) {
       __platform = FlutterMidiCommandLinux();
     } else {
       __platform = MidiCommandPlatform.instance;
     }
-    return __platform;
+    return __platform!;
   }
 
   /// Gets a list of available MIDI devices and returns it.
-  Future<List<MidiDevice>> get devices async {
+  Future<List<MidiDevice>?> get devices async {
     return _platform.devices;
   }
 
@@ -88,14 +87,14 @@ class MidiCommand {
   /// Stream firing events whenever a midi package is received.
   ///
   /// The event contains the raw bytes contained in the MIDI package.
-  Stream<MidiPacket> get onMidiDataReceived {
+  Stream<MidiPacket>? get onMidiDataReceived {
     return _platform.onMidiDataReceived;
   }
 
   /// Stream firing events whenever a change in the MIDI setup occurs.
   ///
   /// For example, when a new BLE devices is discovered.
-  Stream<String> get onMidiSetupChanged {
+  Stream<String>? get onMidiSetupChanged {
     return _platform.onMidiSetupChanged;
   }
 
